backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.115.12
psutil==5.9.6
black==25.1.0
greenlet==3.2.0
apscheduler==3.11.0
h11==0.14.0
idna==3.10
pyotp==2.9.0
qrcode==7.4.2
pillow==10.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-jose==3.4.0
python-multipart==0.0.20
pyinstaller==6.13.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
